Updates table column sizing, moves auxiliary dialogs to new file
